chore(gql): move organization types to organizations subfolder
