+++ app/graphql/mutations/organizations/update.rb
@@ -11,7 +11,7 @@ module Mutations
 
       input_object_class Types::Organizations::UpdateOrganizationInput
 
-      type Types::CurrentOrganizationType
+      type Types::Organizations::CurrentOrganizationType
 
       def resolve(**args)
         result = ::Organizations::UpdateService.call(organization: current_organization, params: args)
