+++ app/graphql/types/payloads/register_user_type.rb
@@ -4,7 +4,7 @@ module Types
   module Payloads
     class RegisterUserType < Types::BaseObject
       field :membership, Types::MembershipType, null: false
-      field :organization, Types::OrganizationType, null: false
+      field :organization, Types::Organizations::OrganizationType, null: false
       field :token, String, null: false
       field :user, Types::UserType, null: false
     end
