+++ app/graphql/types/plans/object.rb
@@ -6,7 +6,7 @@ module Types
       graphql_name 'Plan'
 
       field :id, ID, null: false
-      field :organization, Types::OrganizationType
+      field :organization, Types::Organizations::OrganizationType
 
       field :amount_cents, GraphQL::Types::BigInt, null: false
       field :amount_currency, Types::CurrencyEnum, null: false
