+++ spec/graphql/types/organizations/current_organization_type_spec.rb
@@ -1,8 +1,10 @@
 # frozen_string_literal: true
 
 require 'rails_helper'
-RSpec.describe Types::CurrentOrganizationType do
+RSpec.describe Types::Organizations::CurrentOrganizationType do
   subject { described_class }
+
+  it { is_expected.to be < ::Types::Organizations::BaseOrganizationType }
 
   it { is_expected.to have_field(:id).of_type('ID!') }
