+++ spec/graphql/types/organizations/organization_type_spec.rb
@@ -2,8 +2,10 @@
 
 require 'rails_helper'
 
-RSpec.describe Types::OrganizationType do
+RSpec.describe Types::Organizations::OrganizationType do
   subject { described_class }
+
+  it { is_expected.to be < ::Types::Organizations::BaseOrganizationType }
 
   it { is_expected.to have_field(:id).of_type('ID!') }
   it { is_expected.to have_field(:logo_url).of_type('String') }
